goldfish: refactor goldfish platform configs
On new virtual devices, the goldfish virtual bus can be replaced with
autoprobing infrastructure like Device Tree.  Refactor the goldfish
kernel configs to better accomodate this.

Move the goldfish platform into a menuconfig in the style of the chrome
platform, and separate the goldfish bus into its own config option.

Change-Id: Iaa5b5c9dd4da51b9030229def19b6d7baa36c131
Signed-off-by: Greg Hackmann <ghackmann@google.com>