arm: virt: Support the ranchu machine name as a virt board
The QEMU ranchu board is a variant of the virt board, it is completely
device-tree driven, but should be matched by the virt board.  It
contains a pl011 serial port for console output, but the virt board
doesn't currently initialize the of clock framework, which is required
for this device to be probed, so add this logic to the virt board as
well.

Note that for newer kernels, this has all been reworked to not have a
board file for mach-virt, so it should not be neccessary in that context
and this patch should not be carried foward into newer kernels, but it
is useful for Linux v3.10-based Android kernels.

Change-Id: I250b31c0fc83d02fe06111783f7b36ee1a50abe4
Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>