[MIPS] Ranchu Virtual platform with DT support
Ranchu Virtual platform is a Goldfish based platform
enriched with device tree support. This platform is
backed up by QEMU with the "ranchu" machine which
represents the new virtual board which will be used for
running Android.

The new Ranchu platform introduces support for running Virtio
devices in purpose of reducing the number of Goldfish devices
used by the old Goldfish platform. For example, it uses
virtio-block devices to emulate the Goldfish NAND.

Change-Id: I3a508d0a0ddc3c9d5bf777445ab739f465ab9471
Reviewed-on: https://mipsia.review.mips.com/2859
Reviewed-by: Raghu Gandham <raghu.gandham@imgtec.com>
Tested-by: Raghu Gandham <raghu.gandham@imgtec.com>