MIPS: build fix: sprinkle hardfloat into macros and embedded ASM
Backported commit 8557f958 from android-goldfish-3.10 branch.

GCC before 4.9 haven't passed CFLAGS to the assembler properly,
in this case that flag is -msoft-float.
GCC 4.9 now passes all the CFLAGS to assembler as it should,
but that caused assembler errors in a lot of files:

"Error: opcode not supported on this processor"

This patch fixes it by explicitly setting the hardfloat
attribute in those files.

.set hardfloat

Signed-off-by: Miodrag Dinic <miodrag.dinic@imgtec.com>
Change-Id: Ie2db55c99a63cb8f9168bb30ec731d1df32c2722