MIPS Goldfish: Enable HIGHMEM support
Goldfish platform for MIPS was limited to using up to 496MB
of RAM because the IO space for devices was reserved at
addresses > 496MB.

Using HIGHMEM support, we can map more physical memory without
moving IO space. For this support to work, Goldfish MIPS machine
on the emulator side needs to be updated and 496MB limitation removed.

Change-Id: Ia392000954bacc06f6cdc7b20429e6c9ba15ae36
Signed-off-by: Miodrag Dinic <miodrag.dinic@imgtec.com>