goldfish: enable sw_sync
This CL adds the kernel-space sw_sync module
to the compiled kernel.

This is for implementing EGL sync extensions
so that frames can still be in order but
guest is still allowed to do work.

sw_sync_user can now be disabled, because we
will be using an actual virtual device + driver
to maintain sync timelines instead of guest
userspace simulation of such.

This is part of a sequential, multi-CL change:

external/qemu:

https://android-review.googlesource.com/240119
https://android-review.googlesource.com/239442
https://android-review.googlesource.com/221593
https://android-review.googlesource.com/248563
https://android-review.googlesource.com/248564
https://android-review.googlesource.com/223032

external/qemu-android:

https://android-review.googlesource.com/240155
https://android-review.googlesource.com/238790

kernel/goldfish:

https://android-review.googlesource.com/232631 <- this CL
https://android-review.googlesource.com/238399 <- needs this CL to run

Signed-off-by: Lingfeng Yang <lfy@google.com>
Change-Id: Ia0b9db5c0620b3fb2c87b9ba63a475b694f76442