MIPS: Attach ranchu machine reset and power-off routines
If "reboot" or "reboot -p" are requested from the emulator, these
routines will be triggered which in return should reset or power-off the
machine.

In order to enable this power management device, appropriate device tree node
"goldfish_reset" is expected in the device tree to acquire device memory map
location.

This simple device contains only one register, which holds commands to be performed.
    GORESET: 0x42 command for reboot
    GOPOWEROFF: 0x43 command for shutdown

Change-Id: I151b518a3982cb6c871dedf566846fa07c9304c6
Signed-off-by: Goran Ferenc <goran.ferenc@imgtec.com>